+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 
 ### Fixed
-- Send viewport screenshot on each locators call
+- Send viewport screenshot on each locators call [Trello 1208](https://trello.com/c/tDVd9f7K/1208-visual-locators-returns-empty-locator)
 
 ### Fixed 
-- Send device name when session created
-- Catch NullPointerException if there no visual locators for provided name
+- Send device name when session created [Trello 125](https://trello.com/c/ekZqajRU/1184-device-name-not-set-or-set-to-desktop-when-running-an-appium-test)
+- Catch NullPointerException if there no visual locators for provided name [Trello 1222](https://trello.com/c/jhDnsp9q/1222-java-appium-4-visual-locators-nullpointerexception)
 
 ## [4.2.0] - 2019-10-04
 ### Added