+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-
+## [4.3.0] - 2019-12-23
 ### Fixed
 - Fullpage screenshot with RecyclerView on Android. [Trello 1416](https://trello.com/c/cpHFs1Zr), [Trello 953](https://trello.com/c/3UOEHgCl)
 ### Added